requirements.txt
fastapi
uvicorn
torch
torchaudio
faster-whisper
pyannote.audio
huggingface_hub
ffmpeg-python
imageio-ffmpeg